separate server actions and types for every entity
