+++ src/actions/task/types.ts
@@ -1,0 +1,18 @@
+export type AddTaskParams = {
+  title: string;
+  description?: string;
+};
+
+export type UpdateTaskParams = {
+  id: string;
+  title: string;
+  description?: string;
+};
+
+export type DeleteTaskParams = {
+  id: string;
+};
+
+export type ToggleTaskParams = {
+  id: string;
+};
